requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
pypdf>=4.2
openai>=1.40
python-dotenv>=1.0
tenacity>=8.2
# Only needed for the Stripe webhook handler (src/stripe_webhook.py)
flask>=3.0
stripe>=10.0
